Allow auto-navigation by clicking on log message
This does make the log line a bit harder to read, but it may be
worth it? (and the line was similar to that in Nammu).
It also means we don't store the formatted log, but I think
that's fine, even if we need to get it more than once (not the
case currently).