Read zip files in response correctly
This works but is very ugly. I couldn't find a very easy to deal
with the encoding issues, although this can still be made nicer.

There is something wrong with how we are sending the file content:
the request.log shows a list of errors, and the oracc.log is empty
(because the server couldn't do the validation). It may have to do
with the encoding again; it seems we are sometimes sending more
and sometimes fewer bytes than the server is expecting.