Eliminate some repetition in messages
We can do a lot more, and it may be worth defining a union type
for the request parameters (this would be either a string for the
response ID, or an object containing the command, filename and
file text). We could then have more unified functions which can
distinguish between the two without passing nulls/unhandled values.

(See:
https://www.typescriptlang.org/docs/handbook/2/everyday-types.html#union-types
)