Use correct encoding for zip contents
Converting the zip contents to UTF8 caused the file contents to be
corrupted. The mimemessage package can only convert the message
to a string, so we have to keep the buffer explicitly.

Also, this change correctly replaces *all* line endings with \r\n.
This may not be required, but the previous version only replaced
the first occurrence.