Get validation result correctly (#54)
* Reenable server communication

* Break down for easier changes

* Wait for request status correctly

* [WIP] Get final result from server

* [WIP] Wait for whole response, try to read zips

* Read zip files in response correctly

This works but is very ugly. I couldn't find a very easy to deal
with the encoding issues, although this can still be made nicer.

There is something wrong with how we are sending the file content:
the request.log shows a list of errors, and the oracc.log is empty
(because the server couldn't do the validation). It may have to do
with the encoding again; it seems we are sometimes sending more
and sometimes fewer bytes than the server is expecting.

* Appease the linter and clarify comment

* Send filename, not whole path

Looking at the request.log, we were including the whole path.
This wasn't the source of the errors though.

* Test the final response parsing

* Use an actual log file

* Move response parsing to be with request creation

* Remove unused option

We actually don't send a multipart message when we have a
request ID (although the response to it is multipart).

* Actually remove the unused argument everywhere

* Start switching to node-fetch for HTTP requests

* Don't combine await with chained promises

Apparently mixing the two styles should be avoided because it can
lead to confusion more easily (even though in principle it works).
See for example: https://stackoverflow.com/questions/54385676

* Complete the switch to node-fetch

* Use correct encoding for zip contents

Converting the zip contents to UTF8 caused the file contents to be
corrupted. The mimemessage package can only convert the message
to a string, so we have to keep the buffer explicitly.

Also, this change correctly replaces *all* line endings with \r\n.
This may not be required, but the previous version only replaced
the first occurrence.

* Throw errors more explicitly

* Simplify a little and expand on error messages

The timeout error now matches Nammu.

* Use actual result from server

* Add a test for the initial request creation

* Hide implementation details from validation code

* Add a test for the overall validation function

Primarily to help with refactoring

* Hide more multipart message details in a class

* Test validation of file without errors

Small changes to address some recent strictness by server.
This used (or was supposed to!) to produce no errors.

* Add forgotten file for tests!

* Parse server results correctly on Windows

The server always separates errors with \n. When testing on
Windows, we have to pretend that's the case when we mock the errors
in a file. This is simpler than making sure the file has the right
line endings in the repository for everyone.

* Eliminate some repetition in messages

We can do a lot more, and it may be worth defining a union type
for the request parameters (this would be either a string for the
response ID, or an object containing the command, filename and
file text). We could then have more unified functions which can
distinguish between the two without passing nulls/unhandled values.

(See:
https://www.typescriptlang.org/docs/handbook/2/everyday-types.html#union-types
)

* Fix indentation

* Apply suggestions from review

* Clean up logs a bit

Keeping the errors.
Keeping as info: general progress updates, validation results.
Removing one spurious log, possibly from old debugging.