Activate extension when calling lemmatisation!
Without adding it to the list of activation events, the command
will fail unless the extension has been activated before (e.g if
we click on any of the other buttons first).
Also adding a test to ensure we don't forget to change this if we
add more commands in the future.