Produce CommonJS only for runTests
This requires the experimental "node12" setting for "module",
which is only available on nightly builds.
From trying out different combinations, the "moduleResolution" also
needs to be set to "node12" (this is available for tsc >= 4.5).
It seems that this combination also needs a slight change when
importing default modules ("import * as ... from ..."), and this
requires setting the "esModuleInterop" option (or, more
specifically, the "allowSyntheticDefaultImports" option,
which the former sets).